Added RSA
rsa.py defines an RSA class.
The __init__ method takes two arguments: the maximum modulus that can be generated, and the suggested public key (default 65537). The actual modulus created will be near the maximum, and the public key will be near the suggested key.
Instances of the class have three attributes: n, the modulus; e, the public key; and _d, the private key.
Instances of the class have two methods: encrypt and decrypt.