Added rijndael implementation
rijndael.py implements the rijndael cipher. Keys, plaintext, and ciphertext are input as bitstrings; lengths must be multiples of 32. Nb and Nr can be independently specified, but default to 4 and 6+max(Nr,Nk), respectively. So far, only single-block encrypt and decrypt are implemented. AES128, AES224, AES256 are utility functions that each take a key as an integer argument and return a rijndael instance. rijndael instance encrypt and decrypt methods each take a one-block plaintext input as a bitstring of the appropriate length.
In bitstring.py, allowed bytes argument to __init__, and added string() and bytes() methods to return those arguments, respectively. Note that in python2, bytes doesn't really do the right thing.