Updated signature of bitstring classes
If there are no args, an empty bitstring is returned.
If the (first) arg is an integer, the number of bits is given by the absolute value of the second arg. If the second arg is negative, the bit order of the resulting bitstring is reversed.
If the (first) arg is a bitstring, then if the second arg, the bit order of the resulting bitstring is reversed.
If the (first) arg is a bytes instance or a string, then if the second arg, the octet order of the resulting bitstring is reversed.